Refactor: 클라이밍 상태 업데이트 메서드 수정
